Sorry — I added these but did not commit…Forgot about the two step process
HUGE NUMBER OF CHANGES — all part of the process of moving to a description lattice for managing the creation and indexing of all Krisp individuals.
The basis idea is that an individual, once created, does not change, and has a place in a lattice of descriptions. 
The operation of biding a variable in an existing individual is implemented by creating a new individual that has all the same bindings as the original individual
plus the new one, and which is placed in the description lattice just below the existing individual.

THIS CHANGE HAS BEEN VERIFIED TO PERFORM AS WELL ON ALL OF THE BAISIC CORPORA (NOT THE MITRE ARTICLES) FOR WHICH WE HAVE REGRESSION RESULTS.
IT IS ALSO A BIT FASTER THAN THE ONLD IMPLEMENTATION.