First pass at ASD to load SPARSER.
Relies on an internal-to-SIFT utils system. The load sets the
script var and loads the load-nlp.lisp file. So the classic
load-nlp.lisp approach still works -- and is actually still the
canonical way to load SPARSER.

This ASD provides a convenience mechanism that will make it easier
to define systems that depend on SPARSER.