code to (more) properly capture the set of variables visible at a category (all variables locally defined, plus all variables inherited from super-categories or mixins, except those which are "shadowed" by a variable created by a restriction).
This algorithm is slightly more time consuming than the older, unpredictable one, so we cache the results in a hash table linked to the category by means of another hash-table.

This works at almost exactly the same speed as the old version (maybe slightly faster), and works for biology and blocks-world. It does not, as yet, work for :fire -- that will have to be corrected before this is put on the main branch