re-instate (from deleted code found in git) the mechanisms for contextually interpreting elliptical descriptions. Put it under control of a variable (defparameter *expand-interpretations-in-context* nil   "Turned off because the majority of re-interpretations proposed currently are wrong...")
Need to look at this further -- need to tighten up the cases in which it is run -- in particular, to cases of definite but elliptical reference. This is an interesting question and I believe one that can be solved for a good number of cases. Work to be done over the weekend and next week.