Added "trajectory" category, which composes a direction and a measurement/extent (i.e. "up two steps", "down two octaves", etc.)
For now, it's specifically set to work with two directions ("up" and "down") and pre-defined music units (step, octave), to avoid explicitly defining "step" as a unit of measure and potentially overriding its other meanings, and also the multiple categories currently ascribed to "up" and "down".

In general, though, the idea is that this should just combine any direction and any measurement, and it works nicely as a vp adjunct to movement verbs. Skirts around the need for two separate adjuncts, and I think is semantically sound as a composite category.